prodoctorov/requirements.txt
diskcache
platformdirs
python-telegram-bot